2008-04-16 Roman - !!!:/SYS: Groups of branches control mechanism is changed to direct modification control. - !!!:/DAQ/CIF: Data acquisition of Siemens PLC through ISO_TSAP protocol by library Libnodave is added. The library Libnodave full included to module CIF.     It allow to make support other acquisition methods in future. The module will be renamed to "siemens" in future.     - FIX:/DAQ/Template: Template library's fail elements start error make up is added.     - !:/DAQ/JavaLikeCalc: Build in functions max() and min() is added.     - !:/DAQ/JavaLikeCalc: Symbol '(' for functions control to lexical analisator is added.     - !!!:/UI/QTCfg: Items adding and deleting from nodes tree is added.     - /UI/QTCfg: Nodes adding and deleting watch to nodes tree is added.     - /UI/QTCfg: Input dialog is unified and imported from module Vision.     - /UI/VCAEngin: Widget's identifier length control is added and made it longer - 30 symbols.     - !!:/UI/Vision: Font, Image and Color preview to attributes inspector is added.     - !!:/UI/Vision: Font and Color configure dialog to attribute inspector is added.     - FIX:/UI/Vision, /UI/WebVision: Font name coding and loading is fixed.     - FIX:/UI/Vision, /UI/WebVision: Trend length of view "Trend" of primitive "Diagram" limit of 2300 seconds is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@611 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/WebVision/vca_sess.cpp
@@ -2137,7 +2137,7 @@ void VCADiagram::getReq( SSess &ses )
 	//- Trace cursors value -
 	if( active )
 	{
-	    long long tTimeGrnd = tTime - (int)(tSize*1000000.);
+	    long long tTimeGrnd = tTime - (long long)(tSize*1000000.);
 	    if( curTime >= (tTime-2*(long long)trcPer*1000000) || curTime <= tTimeGrnd )
 		setCursor(tTime,ses.user);
         }
@@ -2148,7 +2148,7 @@ void VCADiagram::getReq( SSess &ses )
 
     //-- Get generic parameters --
     int parNum     = trnds.size();                         		//Parameter's number
-    int tSz        = (int)(tSize*1000000.);       			//Trends size (us)
+    long long tSz  = (long long)(tSize*1000000.);       			//Trends size (us)
     long long tEnd = tTime;                     			//Trends end point (us)
     long long tPict = tEnd;
     long long tBeg = tEnd - tSz;                                      	//Trends begin point (us)
@@ -2594,7 +2594,7 @@ void VCADiagram::setAttrs( XMLNode &node, const string &user )
 
 void VCADiagram::setCursor( long long itm, const string& user )
 {
-    long long tTimeGrnd = tTime - (int)(tSize*1000000.);
+    long long tTimeGrnd = tTime - (long long)(tSize*1000000.);
     curTime = vmax(vmin(itm,tTime),tTimeGrnd);
 
     XMLNode req("set");
@@ -2657,10 +2657,10 @@ void VCADiagram::TrendObj::setAddr( const string &vl )
 						    
 void VCADiagram::TrendObj::loadData( const string &user, bool full )
 {
-    int tSize   = (int)(owner().tSize*1000000.);
+    long long tSize     = (long long)(owner().tSize*1000000.);
     long long tTime     = owner().tTime;
     long long tTimeGrnd = tTime - tSize;
-    int wantPer = tSize/(int)(owner().width+0.5);
+    long long wantPer   = tSize/(int)(owner().width+0.5);
     string arch = owner().valArch;
 		    
     //- Clear trend for empty address and the full reload data -
